+++ rest_framework_mcp/_compat/__init__.py
@@ -1,3 +1,4 @@
 from rest_framework_mcp._compat.acall import acall
+from rest_framework_mcp._compat.reject_awaitable import reject_awaitable
 
-__all__ = ["acall"]
+__all__ = ["acall", "reject_awaitable"]